chore: change lot and bid ids to be uint96

+++ src/interfaces/IHooks.sol
@@ -13,13 +13,13 @@ interface IHooks {
 
     /// @notice Called before payment and payout
     /// TODO define expected state, invariants
-    function pre(uint256 lotId_, uint256 amount_) external;
+    function pre(uint96 lotId_, uint256 amount_) external;
 
     /// @notice Called after payment and before payout
     /// TODO define expected state, invariants
-    function mid(uint256 lotId_, uint256 amount_, uint256 payout_) external;
+    function mid(uint96 lotId_, uint256 amount_, uint256 payout_) external;
 
     /// @notice Called after payment and after payout
     /// TODO define expected state, invariants
-    function post(uint256 lotId_, uint256 payout_) external;
+    function post(uint96 lotId_, uint256 payout_) external;
 }
